Add implementation of tracking summary stats per IP
The new IPStats struct and it's associated bucket (ip_stats) are
going to be used to quickly access summary info about a specific
IP. Stuff that would take a while to crawl. At the moment this
bit of business logic is in the DAL, it would be nice to move it up
a layer (into models).